Add configs for dwm and st
When configuring Suckless programs, it is necessary to compile the
executable from source and modify "config.h". Currently I am playing
around with dwm as a new window manager and st as a new terminal. These
configs should be symlinked to the respective source directories as
"config.h" before building.