Modify tag keys
It is easier to have a custom setup for the tag keys instead of using
actual numbers because of programmer's dvorak keyboard layout. Now most
of the tags have been assigned to adjacent fingers on the right hand.
Additionally a few commands have been defined in preparation for more
shortcut keys.