Move data from rcfiles to vimrc
Although it was nice to have all the different vim settings in seperate,
individual files, practically and with respect to portability it was
more of a hassle. It is easier to just have all settings in the vimrc.
This commit is just moving all the data from the seperate files
to the single vimrc file.