Ignore all file in `.config` and `.local`
A lot of applications create files in `.config` and `.local` (this is
intended because of XDG) which can lead to a lot of extra information
being shown when performing git commands in these directories. This
commit sets git to ignore all the files in these directories so that
only changes for files that have been explicitly added to git will be
displayed.